sqrtf for vector normalize

+++ src/vector.c
@@ -7,7 +7,7 @@
 Vector vector_normalize(Vector v) {
     float mag = (v.x*v.x) + (v.y*v.y) + (v.z*v.z);
     if (fabs(mag - 1.0) > 0.0001) {  // Tolerance.
-        mag = sqrt(mag);
+        mag = sqrtf(mag);
         return (Vector){v.x/mag, v.y/mag, v.z/mag};
     }
     return v;
